fix: bug with pico-UNO response
It was a timing error. The arduino was checking the UART before the pico was ready to send it. Also, there was a UART signal being sent to the UNO at the end of the read data function which the UNO was not reading. This was causing buffer issues which caused the next command to become corrupt